% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain.R
\name{strain}
\alias{strain}
\title{Title}
\usage{
strain(x, column_q, column)
}
\arguments{
\item{x}{A data frame}

\item{column_q}{The column name in "quotes"}

\item{column}{The column name}
}
\description{
Title
}
