% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titlr.R
\name{titlr}
\alias{titlr}
\title{Titlr}
\usage{
titlr(p, title, x.axis, y.axis)
}
\arguments{
\item{p}{Plot name}

\item{title}{"Title text" in quotes}

\item{x.axis}{"X axis text" in quotes}

\item{y.axis}{"Y axis text" in quotes}
}
\value{
p
}
\description{
Titlr
}
\examples{
titlr(plot,"MPG and Weight", "MPG", "Weight")
}
